feat(ime): round20 — recency 近期指数连续化(指数衰减取代五档阶梯)
五档阶梯(10s/1h/5h/1d/3d)同档内同分:两词只差几分钟也无法分辨,
档位边界还有跳变。改为连续时间函数:

  b(t) = min(5, 5 × 2^(−t/半衰期) + (count/3).min(1)),  t>3d → 0(惰性移出)
  半衰期 = 18h(对齐旧阶梯中段锚点:5h≈4.1、1d≈2.2、3d≈0.3)

- 频次增强同步连续化:旧 ≥3 次提交 +1 阶跃 → (count/3) 线性爬升到 +1
- 三层 tier(memory/overlay_dict/wordbook 穿透)u32 → f64;
  pinyin/english 消费点 z=(1-a)(a+b)/8+a 公式不变(b 直用连续值)
- 同步真词头窗口修复(COMPOSE_HEAD_WORDS=10,异步 #38→#6)

验证:218 测试全绿(+2);clippy 0;eval 持平 98.0/99.4;
CLI:提交 nihao→你好,再见时 0.705→0.915(连续 b=5 合成)。

+++ crates/ime-core/src/family/english.rs
@@ -370,11 +370,12 @@ impl EnglishFamily {
         }
         let now = now_ms();
         for c in out.iter_mut() {
-            // 近期指数(时间分档 + 频次增强;三级穿透 L1→L2)。
+            // 近期指数(round20 连续化:指数衰减 + count 线性增强;三级穿透
+            // L1→L2)。
             let b = self.wordbook.tier(&c.text, now);
-            if b > 0 {
+            if b > 0.0 {
                 let a = c.raw_score;
-                c.raw_score = (1.0 - a) * (a + b as f64) / 8.0 + a;
+                c.raw_score = (1.0 - a) * (a + b) / 8.0 + a;
             }
         }
     }
